handling bad requests

+++ server/app.py
@@ -13,6 +13,8 @@ cur = con.cursor()
 def login():
     username = request.args.get("username")
     password = request.args.get("password")
+    if username is None or password is None:
+        return Response(status=400)
     res = cur.execute("SELECT * FROM users WHERE username = ? AND password = ?",(username,password))
     row = res.fetchone()
     if row is None:
@@ -20,8 +22,8 @@ def login():
     else:
         return {"auth_status":True, "user_id":row[0]}
 # {
-#     "auth_status":<bool>
-#     "user_id":<int/null>
+#     "auth_status": <bool>
+#     "user_id": <int/null>
 # }
 # ----------------------------------------------
 
@@ -30,6 +32,8 @@ def login():
 def register():
     username = request.args.get("username")
     password = request.args.get("password")
+    if username is None or password is None:
+        return Response(status=400)
     res = cur.execute("SELECT * FROM users WHERE username = ?",(username,))
     if res.fetchone() is None:
         cur.execute("INSERT INTO users(username,password) VALUES (?,?)",(username,password))
@@ -38,7 +42,7 @@ def register():
     else:
         return {"register_status":False}
 # {
-#     "register_status":<bool>
+#     "register_status": <bool>
 # }
 # ----------------------------------------------
 
@@ -46,9 +50,11 @@ def register():
 @app.route("/post/upvote",methods=["POST"])
 def upvote_post():
     post_id = request.args.get("post_id",type=int)
+    if post_id is None:
+        return Response(status=400)
     cur.execute("UPDATE posts SET upvotes = upvotes + 1 WHERE post_id = ?",(post_id,))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -56,9 +62,11 @@ def upvote_post():
 @app.route("/post/downvote",methods=["POST"])
 def downvote_post():
     post_id = request.args.get("post_id",type=int)
+    if post_id is None:
+        return Response(status=400)
     cur.execute("UPDATE posts SET upvotes = upvotes - 1 WHERE post_id = ?",(post_id,))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -66,9 +74,11 @@ def downvote_post():
 @app.route("/reply/upvote",methods=["POST"])
 def upvote_reply():
     reply_id = request.args.get("reply_id",type=int)
+    if reply_id is None:
+        return Response(status=400)
     cur.execute("UPDATE replies SET upvotes = upvotes + 1 WHERE reply_id = ?",(reply_id,))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -76,9 +86,11 @@ def upvote_reply():
 @app.route("/reply/downvote",methods=["POST"])
 def downvote_reply():
     reply_id = request.args.get("reply_id",type=int)
+    if reply_id is None:
+        return Response(status=400)
     cur.execute("UPDATE replies SET upvotes = upvotes - 1 WHERE reply_id = ?",(reply_id,))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -93,13 +105,19 @@ def downvote_reply():
 # }
 @app.route("/post/create",methods=["POST"])
 def create_post():
+    if request.json is None:
+        return Response(status=400)
     user_id = request.json.get("user_id")
     title = request.json.get("title")
+    if request.json.get("tags") is None:
+        tags = ""
     tags = "\n".join(request.json.get("tags"))
     body = request.json.get("body")
+    if user_id is None or title is None or body is None:
+        return Response(status=400)
     cur.execute("INSERT INTO posts(user_id,title,tags,body) VALUES (?,?,?,?)",(user_id,title,tags,body))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -114,13 +132,19 @@ def create_post():
 # }
 @app.route("/post/edit",methods=["POST"])
 def edit_post():
+    if request.json is None:
+        return Response(status=400)
     post_id = request.json.get("post_id")
     title = request.json.get("title")
+    if request.json.get("tags") is None:
+        tags = ""
     tags = "\n".join(request.json.get("tags"))
     body = request.json.get("body")
+    if post_id is None or title is None or body is None:
+        return Response(status=400)
     cur.execute("UPDATE posts SET title = ?, tags = ?, body = ? WHERE post_id = ?",(title,tags,body,post_id))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -128,10 +152,12 @@ def edit_post():
 @app.route("/post/delete",methods=["POST"])
 def delete_post():
     post_id = request.args.get("post_id",type=int)
+    if post_id is None:
+        return Response(status=400)
     cur.execute("DELETE FROM replies WHERE post_id = ?",(post_id,))
     cur.execute("DELETE FROM posts WHERE post_id = ?",(post_id,))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -143,12 +169,16 @@ def delete_post():
 # }
 @app.route("/reply/create",methods=["POST"])
 def create_reply():
+    if request.json is None:
+        return Response(status=400)
     post_id = request.json.get("post_id")
     user_id = request.json.get("user_id")
     body = request.json.get("body")
+    if post_id is None or user_id is None or body is None:
+        return Response(status=400)
     cur.execute("INSERT INTO replies(post_id,user_id,body) VALUES (?,?,?)",(post_id,user_id,body))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -159,11 +189,15 @@ def create_reply():
 # }
 @app.route("/reply/edit",methods=["POST"])
 def edit_reply():
+    if request.json is None:
+        return Response(status=400)
     reply_id = request.json.get("reply_id")
     body = request.json.get("body")
+    if reply_id is None or body is None:
+        return Response(status=400)
     cur.execute("UPDATE replies SET body = ? WHERE reply_id = ?",(body,reply_id))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
 
@@ -171,12 +205,14 @@ def edit_reply():
 @app.route("/reply/delete",methods=["POST"])
 def delete_reply():
     reply_id = request.args.get("reply_id",type=int)
+    if reply_id is None:
+        return Response(status=400)
     cur.execute("DELETE FROM replies WHERE reply_id = ?",(reply_id,))
     con.commit()
-    return Response(status=204)
+    return Response(status=202)
 # no response body
 # ----------------------------------------------
-# ? posts per page const or variable?
+
 # * /user/posts?user_id=<int>&page=<int>&order_by=<string:[time_asc, time_desc, votes_asc, votes_desc]>
 @app.route("/user/posts",methods=["GET"])
 def get_user_posts():
@@ -184,15 +220,18 @@ def get_user_posts():
     user_id = request.args.get("user_id",type=int)
     order_by = request.args.get("order_by",type=str)
     page = request.args.get("page",type=int)
-    posts = {}
-    res = cur.execute("SELECT COUNT(*) FROM posts WHERE user_id = ?",(user_id,))
-    posts["total_pages"] = math.ceil(res.fetchone()[0]/POSTS_PER_PAGE)
-    posts["posts"] = []
+    if user_id is None or order_by is None or page is None:
+        return Response(status=400)
     if order_by == "time_asc": order_by = "time ASC"
     elif order_by == "time_desc": order_by = "time DESC"
     elif order_by == "votes_asc": order_by = "upvotes ASC"
     elif order_by == "votes_desc": order_by = "upvotes DESC"
+    else: return Response(status=400)
+    posts = {}
+    res = cur.execute("SELECT COUNT(*) FROM posts WHERE user_id = ?",(user_id,))
+    posts["total_pages"] = math.ceil(res.fetchone()[0]/POSTS_PER_PAGE)
     res = cur.execute(f"SELECT * FROM posts WHERE user_id = ? ORDER BY {order_by} LIMIT ? OFFSET ?",(user_id, POSTS_PER_PAGE,(page-1)*POSTS_PER_PAGE))
+    posts["posts"] = []
     for row in res:
         post = {}
         post["post_id"] = row[0]
